+++ eyes.common.java/src/main/java/com/applitools/eyes/SessionAccessibilityStatus.java
@@ -1,0 +1,33 @@
+package com.applitools.eyes;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class SessionAccessibilityStatus {
+    public enum AccessibilityStatus {
+        Passed,
+        Failed
+    }
+
+    @JsonProperty("status")
+    private AccessibilityStatus status;
+    @JsonProperty("level")
+    private AccessibilityLevel level;
+    @JsonProperty("version")
+    private AccessibilityGuidelinesVersion version;
+
+    public SessionAccessibilityStatus() {}
+
+    public SessionAccessibilityStatus(AccessibilityStatus status, AccessibilitySettings settings) {
+        this.status = status;
+        this.level = settings.getLevel();
+        this.version = settings.getGuidelinesVersion();
+    }
+
+    public AccessibilityStatus getStatus() {
+        return status;
+    }
+
+    public AccessibilitySettings getSettings() {
+        return new AccessibilitySettings(level, version);
+    }
+}